// src/Ui.Angular/src/app/shared/generated/model/skill-change-active-request.ts
/**
 * Kaizen.Skills.Service
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: latest
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { SkillSelector } from './skill-selector';

export interface SkillChangeActiveRequest { 
    ToUpdate?: SkillSelector;
    IsActive?: boolean;
}