// src/Ui.Angular/src/app/shared/generated/model/skill-create-request.ts
/**
 * Kaizen.Skills.Service
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: latest
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { SkillCategorySelector } from './skill-category-selector';

export interface SkillCreateRequest { 
    Name?: string;
    ShortDescription?: string;
    Parent?: SkillCategorySelector;
}